LineChart Material, Attempted
I say "attempted" because I was trying to implement a slider system for picking the month ranges.

Turns out the sliders in AngularJS Material are so underdeveloped that a) there are no range sliders, b) it is impossible to use form verification and ng-messages with them, c) it is really hard to make the text boxes as big as is desired.

I will be reverting back to numbers there as soon as possible, but the other integrations went smoothly.