Header + Other Material
These are the last few html pages to get the material upgrade. A merge is coming but I still need to make sure everything is the way I want.

Like icons. I want to add those to the header, actually. You know what, it might be a couple more commits.