Removing Master CLasses
Decided to generalize the insert and update forms to include all of the same things so that two relatively meaningless Java classes can go away.

It also allows instructors more freedom to insert more things at once, which is a good quality of life improvement.

I have a pretty high level of frustration about md-autocomplete, however, as it turns out placing it in an ng-include causes it to massively break in ways that cannot be cured >:(